requirements.txt
bilibili-api-python>=16.0.0
qrcode>=7.3
tqdm>=4.64.0
aiofiles>=23.0.0
